+++ lib/Models/Leaflet.ts
@@ -433,7 +433,7 @@ export default class Leaflet extends GlobeOrMap {
     });
   }
 
-  zoomTo(
+  doZoomTo(
     target:
       | CameraView
       | Rectangle
@@ -478,7 +478,7 @@ export default class Leaflet extends GlobeOrMap {
           }
           if (!isDefined(extent)) {
             // Zoom to the first item!
-            return that.zoomTo(target.mapItems[0], flightDurationSeconds);
+            return that.doZoomTo(target.mapItems[0], flightDurationSeconds);
           }
         } else {
           extent = target.rectangle;