+++ lib/Models/NoViewer.ts
@@ -23,7 +23,7 @@ class NoViewer extends GlobeOrMap {
 
   destroy() {}
 
-  zoomTo(v: CameraView | Rectangle | MappableMixin.MappableMixin, t: any) {
+  doZoomTo(v: CameraView | Rectangle | MappableMixin.MappableMixin, t: any) {
     if (v instanceof CameraView) {
       this._currentView = v;
     } else if (v instanceof Rectangle) {